refactor(fe): [Issue #100-14] 画面 API 直叩きを Hook に移行
History / CategoryManagement / ReceiptScan の API 呼び出しを
features/history・category・receipt の Hook に集約し、Screen を UI 専用に薄型化。
