+++ frontend/src/features/category/index.ts
@@ -1,0 +1,1 @@
+export { useCategoryManagement } from './hooks/useCategoryManagement';
