+++ frontend/src/features/history/index.ts
@@ -1,0 +1,1 @@
+export { useReceiptHistory } from './hooks/useReceiptHistory';
