+++ frontend/src/screens/CategoryManagementScreen.tsx
@@ -1,17 +1,13 @@
-import React, { useState, useEffect } from 'react';
+import React from 'react';
 import { StyleSheet, View, Text, FlatList, ActivityIndicator, Platform } from 'react-native';
-import { categoryApi, type Category } from '../api';
-import { getApiErrorStatus } from '../utils/apiError';
-import { showAlert } from '../utils/alertMessage';
 import { AppBackButton, AppButton, AppListColorDot, AppListItem, AppTextInput } from '../components/ui';
 import { BUTTON_LABELS } from '../constants/buttonLabels';
+import { useCategoryManagement } from '../features/category';
 import { colors } from '../theme/colors';
 import { spacing } from '../theme/spacing';
 import { typography } from '../theme/typography';
 import { cardStyles } from '../theme/cardStyles';
 import { screenLayout } from '../theme/screenLayout';
-import { pickNextCategoryColor } from '../utils/categoryColor';
-import { showConfirmDialog } from '../utils/confirmDialog';
 
 export const CategoryManagementScreen = ({
   onBack,
@@ -20,93 +16,7 @@ export const CategoryManagementScreen = ({
   onBack: () => void;
   currentMemberId: number | null;
 }) => {
-  const [categories, setCategories] = useState<Category[]>([]);
-  const [newName, setNewName] = useState('');
-  const [loading, setLoading] = useState(true);
-  const [optimizing, setOptimizing] = useState(false);
-
-  useEffect(() => {
-    if (currentMemberId) fetchCategories();
-  }, [currentMemberId]);
-
-  const fetchCategories = async () => {
-    if (!currentMemberId) return;
-    setLoading(true);
-    try {
-      const res = await categoryApi.listCategories();
-      setCategories(res.data ?? []);
-    } catch (e: unknown) {
-      if (getApiErrorStatus(e) === 401) {
-        showAlert('セッション切れ', '再度ログインしてください。');
-      } else {
-        showAlert('エラー', 'カテゴリーの取得に失敗しました。');
-      }
-    } finally {
-      setLoading(false);
-    }
-  };
-
-  const addCategory = async () => {
-    if (!newName.trim() || !currentMemberId) return;
-    try {
-      const color = pickNextCategoryColor(
-        categories.map((c) => c.color).filter((c): c is string => !!c)
-      );
-      await categoryApi.createCategory({ name: newName, color });
-      setNewName('');
-      fetchCategories();
-    } catch {
-      showAlert('エラー', '追加に失敗しました。');
-    }
-  };
-
-  const deleteCategory = (id: number) => {
-    showConfirmDialog('削除の確認', 'このカテゴリーを削除しますか？', [
-      { text: 'キャンセル', style: 'cancel' },
-      {
-        text: '削除',
-        style: 'destructive',
-        onPress: async () => {
-          try {
-            await categoryApi.deleteCategory(id);
-            fetchCategories();
-          } catch (e: unknown) {
-            const status = getApiErrorStatus(e);
-            if (status === 400 || status === 409) {
-              showAlert('制限', 'このカテゴリーは既に使用されているため削除できません。');
-            } else {
-              showAlert('エラー', '削除に失敗しました。');
-            }
-          }
-        },
-      },
-    ]);
-  };
-
-  const handleOptimize = () => {
-    showConfirmDialog(
-      'マスタ最適化',
-      'ProductMasterの統計に基づき、カテゴリーのキーワードを自動補強します。よろしいですか？',
-      [
-        { text: 'キャンセル', style: 'cancel' },
-        {
-          text: '実行',
-          onPress: async () => {
-            setOptimizing(true);
-            try {
-              const res = await categoryApi.optimizeCategories();
-              showAlert('完了', res.data?.message ?? '最適化が完了しました。');
-              fetchCategories();
-            } catch {
-              showAlert('エラー', '最適化処理に失敗しました。');
-            } finally {
-              setOptimizing(false);
-            }
-          },
-        },
-      ]
-    );
-  };
+  const category = useCategoryManagement({ currentMemberId });
 
   return (
     <View style={[screenLayout.container, styles.containerCategory]}>
@@ -118,18 +28,18 @@ export const CategoryManagementScreen = ({
       <View style={[cardStyles.section, styles.inputSection]}>
         <AppTextInput
           style={styles.nameInput}
-          value={newName}
-          onChangeText={setNewName}
+          value={category.newName}
+          onChangeText={category.setNewName}
           placeholder="新しいカテゴリー"
         />
-        <AppButton title={BUTTON_LABELS.add} onPress={addCategory} size="md" />
+        <AppButton title={BUTTON_LABELS.add} onPress={category.addCategory} size="md" />
       </View>
 
       <AppButton
         title="🪄 キーワード自動最適化"
-        onPress={handleOptimize}
-        loading={optimizing}
-        disabled={optimizing}
+        onPress={category.handleOptimize}
+        loading={category.optimizing}
+        disabled={category.optimizing}
         fullWidth
         size="md"
         style={{ backgroundColor: colors.semantic.category.optimize, marginBottom: spacing.lg }}
@@ -137,11 +47,11 @@ export const CategoryManagementScreen = ({
 
       {!currentMemberId ? (
         <Text style={styles.emptyText}>メンバーを選択してください</Text>
-      ) : loading ? (
+      ) : category.loading ? (
         <ActivityIndicator size="large" color={colors.primary} style={{ marginTop: 50 }} />
       ) : (
         <FlatList
-          data={categories}
+          data={category.categories}
           keyExtractor={(item) => item.id.toString()}
           contentContainerStyle={styles.list}
           renderItem={({ item }) => (
@@ -151,7 +61,7 @@ export const CategoryManagementScreen = ({
               right={
                 <AppButton
                   title={BUTTON_LABELS.delete}
-                  onPress={() => deleteCategory(item.id)}
+                  onPress={() => category.deleteCategory(item.id)}
                   variant="dangerFilled"
                   size="sm"
                 />
